Add button for adding group and move profile activity
